Add InsertQuery class.
Split code specific to INSERT query generation into a separate class.
