+++ tests/TestCase/Database/Driver/PostgresTest.php
@@ -18,7 +18,7 @@ namespace Cake\Test\TestCase\Database\Driver;
 
 use Cake\Database\Driver\Postgres;
 use Cake\Database\DriverFeatureEnum;
-use Cake\Database\Query;
+use Cake\Database\Query\InsertQuery;
 use Cake\Database\Query\SelectQuery;
 use Cake\Datasource\ConnectionManager;
 use Cake\TestSuite\TestCase;
@@ -161,7 +161,7 @@ class PostgresTest extends TestCase
             ->disableOriginalConstructor()
             ->getMock();
 
-        $query = new Query($connection);
+        $query = new InsertQuery($connection);
         $query->insert(['id', 'title'])
             ->into('articles')
             ->values([1, 'foo']);
@@ -169,7 +169,7 @@ class PostgresTest extends TestCase
         $query = $translator($query);
         $this->assertSame('RETURNING *', $query->clause('epilog'));
 
-        $query = new Query($connection);
+        $query = new InsertQuery($connection);
         $query->insert(['id', 'title'])
             ->into('articles')
             ->values([1, 'foo'])
